+++ relay-server/src/processing/transactions/process.rs
@@ -3,7 +3,8 @@ use std::sync::Arc;
 use relay_base_schema::events::EventType;
 use relay_dynamic_config::ErrorBoundary;
 use relay_event_normalization::GeoIpLookup;
-use relay_event_schema::protocol::{Event, Metrics, SpanV2};
+use relay_event_schema::protocol::{Event, Metrics};
+use relay_profiling::ProfileError;
 use relay_protocol::Annotated;
 use relay_quotas::DataCategory;
 use relay_redis::AsyncRedisClient;
@@ -11,39 +12,41 @@ use relay_sampling::evaluation::{ReservoirEvaluator, SamplingDecision};
 use relay_statsd::metric;
 use smallvec::smallvec;
 
-use crate::envelope::Item;
-use crate::managed::{Counted, Managed, ManagedResult, Quantities, RecordKeeper, Rejected};
+use crate::managed::{Counted, Managed, ManagedResult, Quantities, Rejected};
 use crate::metrics_extraction::transactions::ExtractedMetrics;
+use crate::processing::spans::{Indexed, TotalAndIndexed};
 use crate::processing::transactions::extraction::{self, ExtractMetricsContext};
-use crate::processing::transactions::profile::{Profile, ProfileWithHeaders};
-use crate::processing::transactions::{
-    Error, ExpandedTransaction, ExtractedSpans, Flags, IndexedTransaction, SerializedTransaction,
-    Transaction, TransactionOutput, profile, spans,
-};
+use crate::processing::transactions::spans;
+use crate::processing::transactions::types::{ExpandedTransaction, Flags, Profile};
+use crate::processing::transactions::{Error, SerializedTransaction, profile};
 use crate::processing::utils::event::{
     EventFullyNormalized, EventMetricsExtracted, FiltersStatus, SpansExtracted,
 };
-use crate::processing::{Context, Output, QuotaRateLimiter, utils};
-use crate::services::outcome::{DiscardReason, Outcome};
+use crate::processing::{Context, utils};
+use crate::services::outcome::{DiscardItemType, DiscardReason, Outcome};
 use crate::services::processor::{ProcessingError, ProcessingExtractedMetrics};
 use crate::statsd::{RelayCounters, RelayTimers};
 use crate::utils::SamplingResult;
 
 /// Parses the event payload.
-pub fn parse(
+///
+/// This function boxes the resulting [`ExpandedTransaction`] because keeping it on the stack
+/// causes stack overflows in unit tests when run without optimizations.
+pub fn expand(
     work: Managed<SerializedTransaction>,
-) -> Result<Managed<ExpandedTransaction<Transaction>>, Rejected<Error>> {
-    work.try_map(|work, _| {
+    ctx: Context<'_>,
+) -> Result<Managed<Box<ExpandedTransaction>>, Rejected<Error>> {
+    work.try_map(|work, record_keeper| {
         let SerializedTransaction {
             headers,
-            transaction: transaction_item,
+            event: transaction_item,
             attachments,
-            profile,
+            profiles,
         } = work;
-        let mut transaction = metric!(timer(RelayTimers::EventProcessingDeserialize), {
+        let mut event = metric!(timer(RelayTimers::EventProcessingDeserialize), {
             Annotated::<Event>::from_json_bytes(&transaction_item.payload())
         })?;
-        if let Some(event) = transaction.value_mut() {
+        if let Some(event) = event.value_mut() {
             event.ty = EventType::Transaction.into();
         }
         let flags = Flags {
@@ -51,17 +54,50 @@ pub fn parse(
             spans_extracted: transaction_item.spans_extracted(),
             fully_normalized: headers.meta().request_trust().is_trusted()
                 && transaction_item.fully_normalized(),
+            spans_killswitched: {
+                let span_extraction_sample_rate = ctx
+                    .global_config
+                    .options
+                    .span_extraction_sample_rate
+                    .unwrap_or(1.0);
+                crate::utils::sample(span_extraction_sample_rate).is_discard()
+            },
+            spans_rate_limited: false,
         };
         validate_flags(&flags);
 
-        Ok::<_, Error>(ExpandedTransaction {
+        let mut profiles = profiles.into_iter();
+
+        // Accept at most one profile:
+        let profile = profiles.next();
+        for additional_profile in profiles {
+            record_keeper.reject_err(
+                Outcome::Invalid(DiscardReason::Profiling(relay_profiling::discard_reason(
+                    &ProfileError::TooManyProfiles,
+                ))),
+                additional_profile,
+            );
+        }
+
+        #[cfg(debug_assertions)]
+        {
+            // Fix broken span count headers
+            use relay_protocol::get_value;
+            let embedded = get_value!(event.spans).map_or(0, Vec::len);
+            let diff = embedded as isize - transaction_item.span_count() as isize;
+            record_keeper.modify_by(DataCategory::Span, diff);
+            record_keeper.modify_by(DataCategory::SpanIndexed, diff);
+        }
+
+        Ok::<_, Error>(Box::new(ExpandedTransaction {
             headers,
-            transaction: Transaction(transaction),
+            event,
             flags,
             attachments,
             profile,
-            extracted_spans: ExtractedSpans(vec![]),
-        })
+            extracted_spans: vec![],
+            category: TotalAndIndexed,
+        }))
     })
 }
 
@@ -69,7 +105,6 @@ pub fn parse(
 /// 1. Metrics are only extracted in non-processing relays if the sampling decision is "drop".
 /// 2. That means that if we see a new transaction, it cannot yet have metrics extracted.
 fn validate_flags(flags: &Flags) {
-    debug_assert!(!flags.metrics_extracted);
     if flags.metrics_extracted {
         relay_log::error!("Received a transaction which already had its metrics extracted.");
     }
@@ -77,22 +112,21 @@ fn validate_flags(flags: &Flags) {
 
 /// Validates and massages the data.
 pub fn prepare_data(
-    work: &mut Managed<ExpandedTransaction<Transaction>>,
+    work: &mut Managed<Box<ExpandedTransaction>>,
     ctx: &mut Context<'_>,
     metrics: &mut Metrics,
 ) -> Result<(), Rejected<Error>> {
     let scoping = work.scoping();
     work.try_modify(|work, record_keeper| {
         let profile_id = profile::filter(work, record_keeper, *ctx, scoping.project_id);
-        let event = &mut work.transaction.0;
-        profile::transfer_id(event, profile_id);
-        profile::remove_context_if_rate_limited(event, scoping, *ctx);
+        profile::transfer_id(&mut work.event, profile_id);
+        profile::remove_context_if_rate_limited(&mut work.event, scoping, *ctx);
 
-        utils::dsc::validate_and_set_dsc(&mut work.headers, event, ctx);
+        utils::dsc::validate_and_set_dsc(&mut work.headers, &work.event, ctx);
 
         utils::event::finalize(
             &work.headers,
-            event,
+            &mut work.event,
             work.attachments.iter(),
             metrics,
             ctx.config,
@@ -104,44 +138,121 @@ pub fn prepare_data(
 
 /// Normalizes the transaction event.
 pub fn normalize(
-    work: Managed<ExpandedTransaction<Transaction>>,
+    work: Managed<Box<ExpandedTransaction>>,
     ctx: Context<'_>,
     geoip_lookup: &GeoIpLookup,
-) -> Result<Managed<ExpandedTransaction<Transaction>>, Rejected<Error>> {
+) -> Result<Managed<Box<ExpandedTransaction>>, Rejected<Error>> {
     let project_id = work.scoping().project_id;
-    work.try_map(|mut work, _| {
+    work.try_map(|mut work, r| {
+        let original_span_count = work.count_embedded_spans_and_self();
+
         work.flags.fully_normalized = utils::event::normalize(
             &work.headers,
-            &mut work.transaction.0,
+            &mut work.event,
             EventFullyNormalized(work.flags.fully_normalized),
             project_id,
             ctx,
             geoip_lookup,
         )?
         .0;
+
+        // Normalization may have trimmed spans:
+        debug_assert!(!work.flags.spans_extracted);
+        let new_span_count = work.count_embedded_spans_and_self();
+        if let Some(trimmed) = original_span_count.checked_sub(new_span_count)
+            && trimmed > 0
+        {
+            r.reject_err(
+                Outcome::Invalid(DiscardReason::TooLarge(DiscardItemType::Span)),
+                [
+                    (DataCategory::Span, trimmed),
+                    (DataCategory::SpanIndexed, trimmed),
+                ],
+            );
+        }
+
         Ok::<_, Error>(work)
     })
 }
 
 /// Rejects the entire unit of work if one of the project's filters matches.
 pub fn run_inbound_filters(
-    work: &Managed<ExpandedTransaction<Transaction>>,
+    work: &Managed<Box<ExpandedTransaction>>,
     ctx: Context<'_>,
 ) -> Result<FiltersStatus, Rejected<Error>> {
-    utils::event::filter(&work.headers, &work.transaction.0, &ctx)
+    utils::event::filter(&work.headers, &work.event, &ctx)
         .map_err(ProcessingError::EventFiltered)
         .map_err(Error::from)
         .reject(work)
 }
 
-/// Computes the dynamic sampling decision for the unit of work, but does not perform action on data.
+/// The result of dynamic sampling.
+pub enum SamplingOutput {
+    /// The decision was retain, maintain full transaction.
+    Keep {
+        payload: Managed<Box<ExpandedTransaction>>,
+        sample_rate: Option<f64>,
+    },
+    /// The decision was discard keep only extracted metrics and an optional profile.
+    Drop {
+        metrics: Managed<ExtractedMetrics>,
+        profile: Option<Managed<Profile>>,
+    },
+}
+
+/// Computes the sampling decision for a transaction and associated items.
 pub async fn run_dynamic_sampling(
-    work: &Managed<ExpandedTransaction<Transaction>>,
+    payload: Managed<Box<ExpandedTransaction>>,
+    ctx: Context<'_>,
+    filters_status: FiltersStatus,
+    quotas_client: Option<&AsyncRedisClient>,
+) -> Result<SamplingOutput, Rejected<Error>> {
+    let sampling_result =
+        make_dynamic_sampling_decision(&payload, ctx, filters_status, quotas_client).await;
+
+    let sampling_match = match sampling_result {
+        SamplingResult::Match(m) if m.decision().is_drop() => m,
+        keep => {
+            return Ok(SamplingOutput::Keep {
+                payload,
+                sample_rate: keep.sample_rate(),
+            });
+        }
+    };
+
+    // At this point the decision is to drop the payload.
+    let (payload, metrics) = split_indexed_and_total(payload, ctx, SamplingDecision::Drop)?;
+
+    // Need to process the profile before the event gets dropped:
+    let payload = process_profile(payload, ctx);
+    let (payload, profile) = payload.split_once(|mut payload, _| {
+        let mut profile = payload.profile.take();
+        if let Some(profile) = profile.as_mut() {
+            profile.set_sampled(false);
+        }
+        (payload, profile)
+    });
+
+    let outcome = Outcome::FilteredSampling(sampling_match.into_matched_rules().into());
+    let _ = payload.reject_err(outcome);
+
+    Ok(SamplingOutput::Drop {
+        metrics,
+        profile: profile
+            .transpose()
+            .map(|managed| managed.map(|item, _| Profile(Box::new(item)))),
+    })
+}
+
+/// Computes the dynamic sampling decision for the unit of work, but does not perform action on data.
+async fn make_dynamic_sampling_decision(
+    work: &Managed<Box<ExpandedTransaction>>,
     ctx: Context<'_>,
     filters_status: FiltersStatus,
     quotas_client: Option<&AsyncRedisClient>,
 ) -> SamplingResult {
-    let sampling_result = do_run_dynamic_sampling(work, ctx, filters_status, quotas_client).await;
+    let sampling_result =
+        do_make_dynamic_sampling_decision(work, ctx, filters_status, quotas_client).await;
     relay_statsd::metric!(
         counter(RelayCounters::SamplingDecision) += 1,
         decision = sampling_result.decision().as_str(),
@@ -150,11 +261,11 @@ pub async fn run_dynamic_sampling(
     sampling_result
 }
 
-async fn do_run_dynamic_sampling(
-    work: &Managed<ExpandedTransaction<Transaction>>,
+async fn do_make_dynamic_sampling_decision(
+    work: &Managed<Box<ExpandedTransaction>>,
     ctx: Context<'_>,
     filters_status: FiltersStatus,
-    quotas_client: Option<&AsyncRedisClient>,
+    #[allow(unused)] quotas_client: Option<&AsyncRedisClient>,
 ) -> SamplingResult {
     // Always run dynamic sampling on processing Relays,
     // but delay decision until inbound filters have been fully processed.
@@ -174,46 +285,79 @@ async fn do_run_dynamic_sampling(
     }
     utils::dynamic_sampling::run(
         work.headers.dsc(),
-        work.transaction.0.value(),
+        work.event.value(),
         &ctx,
         Some(&reservoir),
     )
     .await
 }
 
-/// Finishes transaction and profile processing when the dynamic sampling decision was "drop".
-pub fn drop_after_sampling(
-    mut work: Managed<ExpandedTransaction<IndexedTransaction>>,
+type IndexedAndMetrics = (
+    Managed<Box<ExpandedTransaction<Indexed>>>,
+    Managed<ExtractedMetrics>,
+);
+
+/// Splits transaction into indexed payload and metrics representing the total counts.
+pub fn split_indexed_and_total(
+    mut work: Managed<Box<ExpandedTransaction>>,
     ctx: Context<'_>,
-    outcome: Outcome,
-) -> Option<Managed<ProfileWithHeaders>> {
-    work.map(|mut work, record_keeper| {
-        // Take out the profile:
-        let profile = work.profile.take();
-        let headers = work.headers.clone();
+    sampling_decision: SamplingDecision,
+) -> Result<IndexedAndMetrics, Rejected<Error>> {
+    let scoping = work.scoping();
 
-        // reject everything but the profile:
-        record_keeper.reject_err(outcome, work);
+    let had_metrics_extracted = work.flags.metrics_extracted;
+    let extract_span_metrics = !work.flags.spans_rate_limited && !work.flags.spans_killswitched;
 
-        profile.map(|item| ProfileWithHeaders { headers, item })
-    })
-    .transpose()
+    let mut metrics = ProcessingExtractedMetrics::new();
+    work.try_modify(|work, _| {
+        work.flags.metrics_extracted = extraction::extract_metrics(
+            &mut work.event,
+            &mut metrics,
+            ExtractMetricsContext {
+                dsc: work.headers.dsc(),
+                project_id: scoping.project_id,
+                ctx,
+                sampling_decision,
+                metrics_extracted: work.flags.metrics_extracted,
+                extract_span_metrics,
+            },
+        )?
+        .0;
+
+        Ok::<_, Error>(())
+    })?;
+
+    Ok(work.split_once(|work, r| {
+        r.lenient(DataCategory::MetricBucket);
+        if had_metrics_extracted || !work.flags.metrics_extracted {
+            // Invalid config or invalid original transaction
+            r.lenient(DataCategory::Transaction);
+            r.lenient(DataCategory::Span);
+        }
+
+        if work.flags.spans_killswitched {
+            r.lenient(DataCategory::Span);
+        }
+
+        (Box::new(work.into_indexed()), metrics.into_inner())
+    }))
 }
 
 /// Processes the profile attached to the transaction.
-pub fn process_profile(
-    work: Managed<ExpandedTransaction<Transaction>>,
+pub fn process_profile<T>(
+    work: Managed<Box<ExpandedTransaction<T>>>,
     ctx: Context<'_>,
-    sampling_decision: SamplingDecision,
-) -> Managed<ExpandedTransaction<Transaction>> {
+) -> Managed<Box<ExpandedTransaction<T>>>
+where
+    ExpandedTransaction<T>: Counted,
+{
     work.map(|mut work, record_keeper| {
         let mut profile_id = None;
         if let Some(profile) = work.profile.as_mut() {
-            profile.set_sampled(sampling_decision.is_keep());
             let result = profile::process(
                 profile,
                 work.headers.meta().client_addr(),
-                work.transaction.0.value(),
+                work.event.value(),
                 &ctx,
             );
             match result {
@@ -223,60 +367,35 @@ pub fn process_profile(
                 Ok(id) => profile_id = Some(id),
             };
         }
-        profile::transfer_id(&mut work.transaction.0, profile_id);
-        profile::scrub_profiler_id(&mut work.transaction.0);
+        profile::transfer_id(&mut work.event, profile_id);
+        profile::scrub_profiler_id(&mut work.event);
 
         work
     })
 }
 
-type IndexedWithMetrics = (
-    Managed<ExpandedTransaction<IndexedTransaction>>,
-    Managed<ExtractedMetrics>,
-);
-
-/// Extracts transaction & span metrics from the payload.
-pub fn extract_metrics(
-    work: Managed<ExpandedTransaction<Transaction>>,
-    ctx: Context<'_>,
-    sampling_decision: SamplingDecision,
-) -> Result<IndexedWithMetrics, Rejected<Error>> {
-    let project_id = work.scoping().project_id;
-
-    let mut metrics = ProcessingExtractedMetrics::new();
-    let indexed = work.try_map(|mut work, record_keeper| {
-        // Extract metrics here, we're about to drop the event/transaction.
-        work.flags.metrics_extracted = extraction::extract_metrics(
-            &mut work.transaction.0,
-            &mut metrics,
-            ExtractMetricsContext {
-                dsc: work.headers.dsc(),
-                project_id,
-                ctx,
-                sampling_decision,
-                metrics_extracted: work.flags.metrics_extracted,
-                spans_extracted: work.flags.spans_extracted,
-            },
-        )?
-        .0;
-        Ok::<_, Error>(ExpandedTransaction::<IndexedTransaction>::from(work))
-    })?;
-    let metrics = indexed.wrap(metrics.into_inner());
-    Ok((indexed, metrics))
-}
-
 /// Converts the spans embedded in the transaction into top-level span items.
-#[cfg(feature = "processing")]
+///
+/// Only extracts spans in processing.
 pub fn extract_spans(
-    work: Managed<ExpandedTransaction<IndexedTransaction>>,
+    mut work: Managed<Box<ExpandedTransaction>>,
     ctx: Context<'_>,
     server_sample_rate: Option<f64>,
-) -> Managed<ExpandedTransaction<IndexedTransaction>> {
-    work.map(|mut work, r| {
+) -> Managed<Box<ExpandedTransaction>> {
+    if !ctx.is_processing() {
+        return work;
+    }
+
+    work.modify(|work, r| {
+        if work.flags.spans_killswitched {
+            r.lenient(DataCategory::Span);
+            r.lenient(DataCategory::SpanIndexed);
+            return;
+        }
+
         if let Some(results) = spans::extract_from_event(
             work.headers.dsc(),
-            &work.transaction.0,
-            ctx.global_config,
+            &work.event,
             ctx.config,
             server_sample_rate,
             EventMetricsExtracted(work.flags.metrics_extracted),
@@ -285,28 +404,27 @@ pub fn extract_spans(
             work.flags.spans_extracted = true;
             for result in results {
                 match result {
-                    Ok(item) => work.extracted_spans.0.push(item),
-                    Err(_) => r.reject_err(
-                        Outcome::Invalid(DiscardReason::InvalidSpan),
-                        IndexedSpans(1),
-                    ),
+                    Ok(item) => work.extracted_spans.push(item),
+                    Err(_) => {
+                        r.reject_err(
+                            Outcome::Invalid(DiscardReason::InvalidSpan),
+                            IndexedSpans(1),
+                        );
+                    }
                 }
             }
         }
-        work
-    })
+    });
+    work
 }
 
 /// Runs PiiProcessors on the event and its attachments.
-pub fn scrub<T>(
-    work: Managed<ExpandedTransaction<T>>,
+pub fn scrub(
+    work: Managed<Box<ExpandedTransaction>>,
     ctx: Context<'_>,
-) -> Result<Managed<ExpandedTransaction<T>>, Rejected<Error>>
-where
-    T: Counted + AsRef<Annotated<Event>> + AsMut<Annotated<Event>>,
-{
+) -> Result<Managed<Box<ExpandedTransaction>>, Rejected<Error>> {
     work.try_map(|mut work, _| {
-        utils::event::scrub(work.transaction.as_mut(), ctx.project_info)?;
+        utils::event::scrub(&mut work.event, ctx.project_info)?;
         utils::attachments::scrub(work.attachments.iter_mut(), ctx.project_info);
         Ok::<_, Error>(work)
     })